fix: discussionService 파사드 패턴 적용
